+++ frontend/src/metabase/admin/tasks/components/Logs/Logs.unit.spec.js
@@ -1,8 +1,7 @@
 import fetchMock from "fetch-mock";
 import { render, screen } from "@testing-library/react";
-import Logs from "metabase/admin/tasks/containers/Logs";
-
 import { UtilApi } from "metabase/services";
+import { Logs } from "./Logs";
 
 describe("Logs", () => {
   describe("log fetching", () => {
