+++ frontend/src/metabase/admin/tasks/components/Logs/index.ts
@@ -1,0 +1,1 @@
+export * from "./Logs";
